+++ server/lib/requests/subscription.js
@@ -7,15 +7,10 @@ SubscriptionRequest = module.exports = function(session, requestMessage, channel
 
     this.isApproved = false;
 
-    this._timeout = setTimeout(this.decline.bind(this),
-                               this.session.server.options.subscriptionTimeout * 1000);
-
-    util.log('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" started');
+    this.session.server.log('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" started');
 };
 
 SubscriptionRequest.prototype.approve = function() {
-    clearTimeout(this._timeout);
-
     this.isApproved = true;
 
     for (var i = 0; i < this.channels.length; i++) {
@@ -24,21 +19,19 @@ SubscriptionRequest.prototype.approve = function() {
 
     this._sendResponse(true);
 
-    util.log('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" APPROVED');
+    this.session.server.log('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" APPROVED');
 
     //this.session.server.monitor.increment('subscription');
 };
 
 SubscriptionRequest.prototype.decline = function(error) {
-    clearTimeout(this._timeout);
-
     if (this.isApproved) {
         throw new Error('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" already approved');
     }
 
     this._sendResponse(false, error || 'Subscription declined');
 
-    util.log('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" DECLINED' + (error ? ': ' + error : ''));
+    this.session.server.log('Session ' + this.session.id + ' subscription request to channel "' + this._getChannelNames() + '" DECLINED' + (error ? ': ' + error : ''));
 
     //this.session.server.monitor.increment('declinedSubscription');
 };