+++ server/lib/requests/unsubscription.js
@@ -7,15 +7,10 @@ UnsubscriptionRequest = module.exports = function(session, requestMessage, chann
 
     this.isApproved = false;
 
-    this._timeout = setTimeout(this.decline.bind(this),
-                               this.session.server.options.unsubscriptionTimeout * 1000);
-
-    util.log('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" started');
+    this.session.server.log('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" started');
 };
 
 UnsubscriptionRequest.prototype.approve = function() {
-    clearTimeout(this._timeout);
-
     this.isApproved = true;
 
     for (var i = 0; i < this.channels.length; i++) {
@@ -24,19 +19,17 @@ UnsubscriptionRequest.prototype.approve = function() {
 
     this._sendResponse(true);
 
-    util.log('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" APPROVED');
+    this.session.server.log('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" APPROVED');
 };
 
 UnsubscriptionRequest.prototype.decline = function(error) {
-    clearTimeout(this._timeout);
-
     if (this.isApproved) {
         throw new Error('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" already approved');
     }
 
     this._sendResponse(false, error || 'Unsubscription declined');
 
-    util.log('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" DECLINED' + (error ? ': ' + error : ''));
+    this.session.server.log('Session ' + this.session.id + ' unsubscription request to channel "' + this._getChannelNames() + '" DECLINED' + (error ? ': ' + error : ''));
 };
 
 UnsubscriptionRequest.prototype._sendResponse = function(successful, error) {